+++ ouroboros-network-framework/sim-tests/Test/Simulation/Network/Snocket.hs
@@ -218,7 +218,7 @@ clientServerSimulation payloads =
                   listen snocket fd
                   accept snocket fd >>= acceptLoop threadsVar)
           `finally` do
-            threads <- atomically (readTVar threadsVar)
+            threads <- readTVarIO threadsVar
             traverse_ cancel threads
       where
         acceptLoop :: StrictTVar m (Set (Async m ()))
@@ -242,13 +242,12 @@ clientServerSimulation payloads =
         handleConnection bearer remoteAddr = do
           labelThisThread "server-handler"
           bracket
-            (newMux (MiniProtocolBundle
-                      [ MiniProtocolInfo {
-                            miniProtocolNum    = reqRespProtocolNum,
-                            miniProtocolDir    = ResponderDirectionOnly,
-                            miniProtocolLimits = MiniProtocolLimits maxBound
-                          }
-                      ]))
+            (newMux [ MiniProtocolInfo {
+                        miniProtocolNum    = reqRespProtocolNum,
+                        miniProtocolDir    = ResponderDirectionOnly,
+                        miniProtocolLimits = MiniProtocolLimits maxBound
+                      }
+                    ])
             stopMux
             $ \mux -> do
               let connId = ConnectionId {
@@ -286,13 +285,12 @@ clientServerSimulation payloads =
                 (close snocket)
                 $ \fd -> do
                   connect snocket fd serverAddr
-                  mux <- newMux (MiniProtocolBundle
-                                  [ MiniProtocolInfo {
-                                        miniProtocolNum    = reqRespProtocolNum,
-                                        miniProtocolDir    = InitiatorDirectionOnly,
-                                        miniProtocolLimits = MiniProtocolLimits maxBound
-                                      }
-                                  ])
+                  mux <- newMux [ MiniProtocolInfo {
+                                    miniProtocolNum    = reqRespProtocolNum,
+                                    miniProtocolDir    = InitiatorDirectionOnly,
+                                    miniProtocolLimits = MiniProtocolLimits maxBound
+                                  }
+                                ]
                   localAddr <- getLocalAddr snocket fd
                   let connId = ConnectionId {
                           localAddress  = localAddr,